Set up treemacs
In order to get this set up, I needed to make sure that version 1.12.0
of s.el was installed because version 1.11.0 of s.el did not include the
function s-blank-str?, which is used by treemacs. Since the elpy package
archive took higher priority than either melpa or melpa-stable, that was
where s.el was being installed from, since it happened to have a copy of
s.el, and this copy was only version 1.11.0.

It now seems that elpy is distributed on both melpa-stable and melpa, so
there is no reason to have the elpy package archive included in the
setup.

For reference, see:

- https://github.com/Groovy-Emacs-Modes/groovy-emacs-modes/issues/96
- https://github.com/magnars/s.el/issues/114