Add forge from melpa and pin magit to melpa
There is a newish package forge for interacting with Git forges such as
GitHub and GitLab. There seem to be some bugs when installed alongside
the magit version from MELPA Stable, so I've pinned magit to MELPA.

This newer version of magit replaced magit-popup with transient, which
required changing a few things. Default arguments to commands like log
and commit are now stored inside of transient/values.el. Moreover, there
is another file, transient/levels.el, which handles which options are
actually displayed in the popup. You can set the level for a particular
transient option to be either 1 or 7. By default, things equal to or
less than 4 show up.